test_words_api/requirements.txt
Django>=4.2.7
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.1
django-simpleui>=2023.12.12
django-redis>=5.4.0
mysqlclient>=2.2.0
gunicorn>=21.2.0
celery>=5.3.4
redis>=5.0.1
Pillow>=10.1.0
python-decouple>=3.8
django-filter>=23.5
drf-yasg>=1.21.7 